+++ Podfile
@@ -11,9 +11,9 @@ use_frameworks!
 # - `{ {kit spec hash} => {sdk spec hash}` to depend on specific pod options (:git => …, :podspec => …) for each repo. Used by Fastfile during CI
 #
 # Warning: our internal tooling depends on the name of this variable name, so be sure not to change it
-$matrixKitVersion = '= 0.15.7'
+# $matrixKitVersion = '= 0.15.7'
 # $matrixKitVersion = :local
-# $matrixKitVersion = {'develop' => 'develop'}
+$matrixKitVersion = {'develop' => 'develop'}
 
 ########################################
 
@@ -46,7 +46,7 @@ abstract_target 'RiotPods' do
   pod 'GBDeviceInfo', '~> 6.6.0'
   pod 'Reusable', '~> 4.1'
   pod 'KeychainAccess', '~> 4.2.2'
- 
+
   # Piwik for analytics
   pod 'MatomoTracker', '~> 7.4.1'
 
@@ -103,7 +103,7 @@ post_install do |installer|
       # Plus the app does not enable it
       config.build_settings['ENABLE_BITCODE'] = 'NO'
 
-      # Make fastlane(xcodebuild) happy by preventing it from building for arm64 simulator 
+      # Make fastlane(xcodebuild) happy by preventing it from building for arm64 simulator
       config.build_settings["EXCLUDED_ARCHS[sdk=iphonesimulator*]"] = "arm64"
 
       # Force ReadMoreTextView to use Swift 5.2 version (as there is no code changes to perform)
@@ -111,7 +111,7 @@ post_install do |installer|
         config.build_settings['SWIFT_VERSION'] = '5.2'
       end
 
-      # Stop Xcode 12 complaining about old IPHONEOS_DEPLOYMENT_TARGET from pods 
+      # Stop Xcode 12 complaining about old IPHONEOS_DEPLOYMENT_TARGET from pods
       config.build_settings.delete 'IPHONEOS_DEPLOYMENT_TARGET'
     end
   end
