+++ RiotSwiftUI/Modules/Room/NotificationSettings/Model/RoomNotificationSettingsViewState.swift
@@ -1,0 +1,43 @@
+// 
+// Copyright 2021 New Vector Ltd
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+import Foundation
+
+/// RoomNotificationSettingsViewController view state
+struct RoomNotificationSettingsViewState: RoomNotificationSettingsViewStateType {
+    let roomEncrypted: Bool
+    var saving: Bool
+    var notificationState: RoomNotificationState
+    var avatarData: AvatarType?
+    var displayName: String?
+}
+
+extension RoomNotificationSettingsViewState {
+    var notificationOptions: [RoomNotificationState] {
+        if roomEncrypted {
+            return [.all, .mute]
+        } else {
+            return RoomNotificationState.allCases
+        }
+    }
+}
+
+extension RoomNotificationSettingsViewState {
+    var roomEncryptedString: String {
+        roomEncrypted ? VectorL10n.roomNotifsSettingsEncryptedRoomNotice : ""
+    }
+}
+
